+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.15 - 2023-04-05
+
+### Added
+* New setting `turnAngle` to set the angle per key press for the WSAD controller.
+
+### Removed
+* Command `turn-angle`.
+
 ## 0.0.14 - 2023-04-05
 
 ### Removed